changing user tests location
