+++ posts/tests/tests.py
@@ -49,51 +49,6 @@ class PostEditTestCase(TestCase):
         self.client.post(reverse("post_edit", kwargs={"pk": self.post.pk}), data=post_data)
 
 
-class UserAuthenticationTestCase(TestCase):
-    def setUp(self):
-        self.client = Client()
-
-    def user_creation(self):
-        self.user = User.objects.create_user(username="testuser", email="test@gmail.com", password="password")
-        found_user = User.objects.filter(username="testuser", email="test@gmail.com", password="password")
-        self.assertEqual(found_user.count(), 1)
-
-    def test_user_registration(self):
-        # Make a POST request to register a new user
-        response = self.client.post(
-            reverse("register"),
-            {
-                "username": "testuser",
-                "email": "test@gmail.com",
-                "password1": "password",
-                "password2": "password",
-            },
-        )
-
-        # Check that the user was created successfully
-        self.assertEqual(response.status_code, 200)
-        found_user = User.objects.filter(username="testuser", email="test@gmail.com")
-        self.assertEqual(found_user.count(), 1)
-
-    def test_user_login(self):
-        # Create a test user
-        self.user = User.objects.create_user(username="testuser", email="test@gmail.com", password="password")
-
-        # Make a POST request to log in as the test user
-        response = self.client.post(
-            reverse("login"),
-            {
-                "username": "testuser",
-                "password": "password",
-            },
-        )
-
-        # Check that the login was successful
-        self.assertEqual(response.status_code, 302)
-        self.assertTrue("_auth_user_id" in self.client.session)
-        self.assertEqual(int(self.client.session["_auth_user_id"]), self.user.id)
-
-
 class BadgeTestCase(TestCase):
     def setUp(self):
         self.client = Client()
